a function's canvas must be in its component
